refactor: Apply updated paths for the view and logic files
