+++ src/components/cards/CardItem/index.jsx
@@ -7,8 +7,8 @@ import CardTags from '@/components/common/CardTags';
 import CommonModal from '@/components/layout/modal/CommonModal';
 import DetailCard from '@/components/dashboard/modal/card/DetailCard';
 import IconModal from '@/components/layout/modal/IconModal';
-import useCardFetchData from '@/hooks/logic/useCardFetchData';
-import useDeleteCard from '@/hooks/logic/useDeleteCard';
+import useCardFetchData from '@/hooks/card/useCardFetchData';
+import useDeleteCard from '@/hooks/card/useDeleteCard';
 import useModalState from '@/hooks/useModalState';
 import { IMAGE_REGEX, ICON } from '@/constants';
 import styles from './CardItem.module.scss';
